requirements.txt
MySQL-python
peewee
bottle

### paho-mqtt

# https://github.com/gerold-penz/python-jsonrpc
python-jsonrpc

docopt==0.6.1
pynacl
